requirements.txt
langchain>=0.1.0
langchain-community
langchain-openai
openai>=1
arize-phoenix[evals]
tiktoken
openinference-instrumentation-openai
streamlit==1.39.0
pypdf
streamlit_chat
boto3
google-cloud-bigquery
google-cloud-aiplatform
openinference-instrumentation-bedrock
